villucheckdelete

+++ Vln-2016/consoleui.cpp
@@ -274,11 +274,13 @@ void consoleUI::run()
         //Delete from database
         listServices scientist;
         int ID = 0;
+        bool personInDatabase = true;
+        string doubleCheck;
         string name;
         cout << "Enter name: ";
         cin >> name;
         print();
-        for(int i = 0 ; i < scientist.getSize();)
+        for(int i = 0 ; i < scientist.getSize(); i++)
         {
             if(name ==  scientist.changeToLower(scientist.getFirstNameFromList(i)) || name == scientist.changeToLower(scientist.getLastNameFromList(i)))
             {
@@ -318,26 +320,44 @@ void consoleUI::run()
                 cout << scientist.getAwardsFromList(i) << left;;
                 cout << "   *" << endl;
 
-
-
-                i++;
             }
-            else
+            else if(name !=  scientist.changeToLower(scientist.getFirstNameFromList(i)) || name != scientist.changeToLower(scientist.getLastNameFromList(i)))
             {
-                i++;
+                if(i == scientist.getSize() - 1)
+                {
+                    cout << "Person is not in database!" << endl;
+                    personInDatabase = false;
+                    break;
+                }
             }
         }
-        for(int i = 0 ; i < 9 ; i++)
+        for(int i = 0 ; i < 8 ; i++)
         {
             cout << "--------";
+        }
+        if(personInDatabase == false)
+        {
+            break;
         }
         cout << endl;
         cout << "Enter number for scientist you want to delete: ";
         cin >> ID;
-        scientist.deleteFromList(ID);
-        print();
-        printNames(scientist);
-        scientist.writeNewFile();
+        cout << "Are you sure you want to delete ?: " << endl;
+        cout << "write Yes to delete or any character to abort: "<< endl;
+        cin >> doubleCheck;
+        scientist.changeToLower(doubleCheck);
+        if(doubleCheck == "yes")
+        {
+            scientist.deleteFromList(ID);
+            print();
+            printNames(scientist);
+            scientist.writeNewFile();
+        }
+        else
+        {
+            break;
+        }
+
     }
         break;
  }
